pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "machine"
version = "0.1"
requires-python = ">=3.14"
dependencies = [
  "anthropic",
  "google-api-core",
  "google-genai",
  "langcodes",
  "openai",
  "tenacity",
  "typer",
]
license = "MIT"
license-files = ["LICENSE"]

[project.optional-dependencies]
dev = ["pytest-sugar"]
test = ["pytest", "pytest-asyncio", "pytest-random-order", "pytest-cov"]

[project.scripts]
transpile = "machine.transpile:main"


[tool.coverage.run]
branch = true
data_file = "tests/.coverage"
parallel = true
relative_files = true
source = ["."]

[tool.coverage.report]
exclude_lines = ["@overload", "if TYPE_CHECKING:", "pragma: no cover"]
fail_under = 100

[tool.coverage.html]
directory = "tests/coverage-html"
show_contexts = true

[tool.coverage.xml]
output = "tests/coverage.xml"


[tool.hatch.build.targets.wheel]
packages = ["src/machine"]

[tool.hatch.build.targets.wheel.force-include]
"languages.md" = "machine/languages.md"
"machine.md" = "machine/machine.md"


[tool.pytest.ini_options]
addopts = """
  --showlocals
  --tb=short
"""
asyncio_mode = "auto"
cache_dir = "tests/.cache"
tmp_path_retention_policy = "failed"
